Extract scroll functionality into ScrollGroup
