+++ src/main/java/com/maddyhome/idea/vim/group/MotionGroup.java
@@ -8,7 +8,10 @@
 package com.maddyhome.idea.vim.group;
 
 import com.intellij.openapi.actionSystem.DataContext;
-import com.intellij.openapi.editor.*;
+import com.intellij.openapi.editor.Caret;
+import com.intellij.openapi.editor.Editor;
+import com.intellij.openapi.editor.LogicalPosition;
+import com.intellij.openapi.editor.VisualPosition;
 import com.intellij.openapi.fileEditor.FileEditor;
 import com.intellij.openapi.fileEditor.FileEditorManagerEvent;
 import com.intellij.openapi.fileEditor.TextEditor;
@@ -38,19 +41,15 @@ import com.maddyhome.idea.vim.mark.Mark;
 import com.maddyhome.idea.vim.newapi.IjExecutionContext;
 import com.maddyhome.idea.vim.newapi.IjVimCaret;
 import com.maddyhome.idea.vim.newapi.IjVimEditor;
-import com.maddyhome.idea.vim.options.LocalOptionChangeListener;
 import com.maddyhome.idea.vim.options.OptionConstants;
 import com.maddyhome.idea.vim.options.OptionScope;
 import com.maddyhome.idea.vim.ui.ex.ExEntryPanel;
-import com.maddyhome.idea.vim.vimscript.model.datatypes.VimDataType;
 import com.maddyhome.idea.vim.vimscript.model.datatypes.VimInt;
 import kotlin.Pair;
-import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.Range;
 
-import java.awt.*;
 import java.io.File;
 
 import static com.maddyhome.idea.vim.group.ChangeGroup.*;
@@ -149,7 +148,7 @@ public class MotionGroup extends VimMotionGroupBase {
     return new TextRange(start, end);
   }
 
-  private static void moveCaretToView(@NotNull Editor editor) {
+  public static void moveCaretToView(@NotNull Editor editor) {
     final int scrollOffset = getNormalizedScrollOffset(editor);
 
     final int topVisualLine = getVisualLineAtTopOfScreen(editor);
@@ -203,35 +202,6 @@ public class MotionGroup extends VimMotionGroupBase {
 
       UserDataManager.setVimLastColumn(editor.getCaretModel().getPrimaryCaret(), col);
     }
-  }
-
-  // Get the visual line that will be in the same screen relative location as the current caret line, after the screen
-  // has been scrolled
-  private static int getScrollScreenTargetCaretVisualLine(final @NotNull Editor editor, int rawCount, boolean down) {
-    final Rectangle visibleArea = getVisibleArea(editor);
-    final int caretVisualLine = editor.getCaretModel().getVisualPosition().line;
-    final int scrollOption = getScrollOption(rawCount);
-
-    int targetCaretVisualLine;
-    if (scrollOption == 0) {
-      // Scroll up/down half window size by default. We can't use line count here because of block inlays
-      final int offset = down ? (visibleArea.height / 2) : editor.getLineHeight() - (visibleArea.height / 2);
-      targetCaretVisualLine = editor.yToVisualLine(editor.visualLineToY(caretVisualLine) + offset);
-    }
-    else {
-      targetCaretVisualLine = down ? caretVisualLine + scrollOption : caretVisualLine - scrollOption;
-    }
-
-    return EngineEditorHelperKt.normalizeVisualLine(new IjVimEditor(editor), targetCaretVisualLine);
-  }
-
-  private static int getScrollOption(int rawCount) {
-    if (rawCount == 0) {
-      return ((VimInt) VimPlugin.getOptionService().getOptionValue(OptionScope.GLOBAL.INSTANCE, OptionConstants.scrollName, OptionConstants.scrollName)).getValue();
-    }
-    // TODO: This needs to be reset whenever the window size changes
-    VimPlugin.getOptionService().setOptionValue(OptionScope.GLOBAL.INSTANCE, OptionConstants.scrollName, new VimInt(rawCount), OptionConstants.scrollName);
-    return rawCount;
   }
 
   private static int getNormalizedScrollOffset(final @NotNull Editor editor) {
@@ -317,52 +287,6 @@ public class MotionGroup extends VimMotionGroupBase {
   }
 
   @Override
-  public boolean scrollCurrentLineToDisplayTop(@NotNull VimEditor editor, int rawCount, boolean start) {
-    scrollLineToScreenLocation(((IjVimEditor)editor).getEditor(), ScreenLocation.TOP, rawCount, start);
-    return true;
-  }
-
-  @Override
-  public boolean scrollCurrentLineToDisplayMiddle(@NotNull VimEditor editor, int rawCount, boolean start) {
-    scrollLineToScreenLocation(((IjVimEditor)editor).getEditor(), ScreenLocation.MIDDLE, rawCount, start);
-    return true;
-  }
-
-  @Override
-  public boolean scrollCurrentLineToDisplayBottom(@NotNull VimEditor editor, int rawCount, boolean start) {
-    scrollLineToScreenLocation(((IjVimEditor)editor).getEditor(), ScreenLocation.BOTTOM, rawCount, start);
-    return true;
-  }
-
-  @Override
-  public boolean scrollCaretColumnToDisplayLeftEdge(@NotNull VimEditor vimEditor) {
-    Editor editor = ((IjVimEditor)vimEditor).getEditor();
-    final VisualPosition caretVisualPosition = editor.getCaretModel().getVisualPosition();
-    final int scrollOffset = getNormalizedSideScrollOffset(editor);
-    // TODO: Should the offset be applied to visual columns? This includes inline inlays and folds
-    final int column = max(0, caretVisualPosition.column - scrollOffset);
-    scrollColumnToLeftOfScreen(editor, caretVisualPosition.line, column);
-    return true;
-  }
-
-  @Override
-  public boolean scrollCaretColumnToDisplayRightEdge(@NotNull VimEditor editor) {
-    Editor ijEditor = ((IjVimEditor)editor).getEditor();
-    final VisualPosition caretVisualPosition = ijEditor.getCaretModel().getVisualPosition();
-    final int scrollOffset = getNormalizedSideScrollOffset(ijEditor);
-    // TODO: Should the offset be applied to visual columns? This includes inline inlays and folds
-    final int column =
-      EngineEditorHelperKt.normalizeVisualColumn(editor, caretVisualPosition.line, caretVisualPosition.column + scrollOffset, false);
-    scrollColumnToRightOfScreen(ijEditor, caretVisualPosition.line, column);
-    return true;
-  }
-
-  @Override
-  public void scrollCaretIntoView(@NotNull VimEditor editor) {
-    ScrollViewHelper.scrollCaretIntoView(((IjVimEditor) editor).getEditor());
-  }
-
-  @Override
   public int moveCaretToFirstDisplayLine(@NotNull VimEditor editor,
                                          @NotNull ImmutableVimCaret caret,
                                          int count,
@@ -384,25 +308,6 @@ public class MotionGroup extends VimMotionGroupBase {
   public int moveCaretToMiddleDisplayLine(@NotNull VimEditor editor, @NotNull ImmutableVimCaret caret) {
     return moveCaretToScreenLocation(((IjVimEditor)editor).getEditor(), ((IjVimCaret)caret).getCaret(),
                                      ScreenLocation.MIDDLE, 0, false);
-  }
-
-  @Override
-  public boolean scrollLines(@NotNull VimEditor editor, int lines) {
-    assert lines != 0 : "lines cannot be 0";
-    Editor ijEditor = ((IjVimEditor)editor).getEditor();
-
-    if (lines > 0) {
-      final int visualLine = getVisualLineAtTopOfScreen(ijEditor);
-      scrollVisualLineToTopOfScreen(ijEditor, visualLine + lines);
-    }
-    else {
-      final int visualLine = getNonNormalizedVisualLineAtBottomOfScreen(ijEditor);
-      scrollVisualLineToBottomOfScreen(ijEditor, visualLine + lines);
-    }
-
-    moveCaretToView(ijEditor);
-
-    return true;
   }
 
   @Override
@@ -507,38 +412,6 @@ public class MotionGroup extends VimMotionGroupBase {
     return new Motion.AbsoluteOffset(offset);
   }
 
-
-  @Override
-  public boolean scrollColumns(@NotNull VimEditor editor, int columns) {
-    Editor ijEditor = ((IjVimEditor)editor).getEditor();
-    final VisualPosition caretVisualPosition = ijEditor.getCaretModel().getVisualPosition();
-    if (columns > 0) {
-      // TODO: Don't add columns to visual position. This includes inlays and folds
-      int visualColumn = EngineEditorHelperKt.normalizeVisualColumn(editor, caretVisualPosition.line,
-                                               getVisualColumnAtLeftOfDisplay(ijEditor, caretVisualPosition.line) +
-                                               columns, false);
-
-      // If the target column has an inlay preceding it, move passed it. This inlay will have been (incorrectly)
-      // included in the simple visual position, so it's ok to step over. If we don't do this, scrollColumnToLeftOfScreen
-      // can get stuck trying to make sure the inlay is visible.
-      // A better solution is to not use VisualPosition everywhere, especially for arithmetic
-      final Inlay<?> inlay =
-        ijEditor.getInlayModel().getInlineElementAt(new VisualPosition(caretVisualPosition.line, visualColumn - 1));
-      if (inlay != null && !inlay.isRelatedToPrecedingText()) {
-        visualColumn++;
-      }
-
-      scrollColumnToLeftOfScreen(ijEditor, caretVisualPosition.line, visualColumn);
-    }
-    else {
-      // Don't normalise the rightmost column, or we break virtual space
-      final int visualColumn = getVisualColumnAtRightOfDisplay(ijEditor, caretVisualPosition.line) + columns;
-      scrollColumnToRightOfScreen(ijEditor, caretVisualPosition.line, visualColumn);
-    }
-    moveCaretToView(ijEditor);
-    return true;
-  }
-
   @Override
   public @NotNull Motion moveCaretToCurrentDisplayLineStart(@NotNull VimEditor editor, ImmutableVimCaret caret) {
     final int col =
@@ -561,58 +434,6 @@ public class MotionGroup extends VimMotionGroupBase {
     final int col =
       getVisualColumnAtRightOfDisplay(((IjVimEditor)editor).getEditor(), caret.getVisualPosition().getLine());
     return moveCaretToColumn(editor, caret, col, allowEnd);
-  }
-
-  @Override
-  public boolean scrollFullPageDown(@NotNull VimEditor editor, @NotNull VimCaret caret, int pages) {
-    Editor ijEditor = ((IjVimEditor)editor).getEditor();
-    Caret ijCaret = ((IjVimCaret)caret).getCaret();
-    final Pair<Boolean, Integer> result = EditorHelper.scrollFullPageDown(ijEditor, pages);
-
-    final int scrollOffset = getNormalizedScrollOffset(ijEditor);
-    final int topVisualLine = getVisualLineAtTopOfScreen(ijEditor);
-    int caretVisualLine = result.getSecond();
-    if (caretVisualLine < topVisualLine + scrollOffset) {
-      caretVisualLine =
-        EngineEditorHelperKt.normalizeVisualLine(new IjVimEditor(ijEditor), caretVisualLine + scrollOffset);
-    }
-
-    if (caretVisualLine != ijCaret.getVisualPosition().line) {
-      final int offset =
-        moveCaretToLineWithStartOfLineOption(editor, EngineEditorHelperKt.visualLineToBufferLine(editor, caretVisualLine), caret);
-      caret.moveToOffset(offset);
-      return result.getFirst();
-    }
-
-    return false;
-  }
-
-  @Override
-  public boolean scrollFullPageUp(@NotNull VimEditor editor, @NotNull VimCaret caret, int pages) {
-    Editor ijEditor = ((IjVimEditor)editor).getEditor();
-    Caret ijCaret = ((IjVimCaret)caret).getCaret();
-    final Pair<Boolean, Integer> result = EditorHelper.scrollFullPageUp(ijEditor, pages);
-
-    final int scrollOffset = getNormalizedScrollOffset(ijEditor);
-    final int bottomVisualLine = getVisualLineAtBottomOfScreen(ijEditor);
-    int caretVisualLine = result.getSecond();
-    if (caretVisualLine > bottomVisualLine - scrollOffset) {
-      caretVisualLine =
-        EngineEditorHelperKt.normalizeVisualLine(new IjVimEditor(ijEditor), caretVisualLine - scrollOffset);
-    }
-
-    if (caretVisualLine != ijCaret.getVisualPosition().line && caretVisualLine != -1) {
-      final int offset =
-        moveCaretToLineWithStartOfLineOption(editor, EngineEditorHelperKt.visualLineToBufferLine(editor, caretVisualLine), caret);
-      caret.moveToOffset(offset);
-      return result.getFirst();
-    }
-
-    // We normally error if we didn't move the caret, but we have a special case for a page showing only the last two
-    // lines of the file and virtual space. Vim normally scrolls window height minus two, but when the caret is on last
-    // line minus one, this becomes window height minus one, meaning the top line of the current page becomes the bottom
-    // line of the new page, and the caret doesn't move. Make sure we don't beep in this scenario.
-    return caretVisualLine == EngineEditorHelperKt.getVisualLineCount(editor) - 2;
   }
 
   public @Range(from = 0, to = Integer.MAX_VALUE) int moveCaretToLineWithSameColumn(@NotNull VimEditor editor,
@@ -643,109 +464,6 @@ public class MotionGroup extends VimMotionGroupBase {
     }
     else {
       return moveCaretToLineWithSameColumn(editor, line, caret);
-    }
-  }
-
-  @Override
-  public boolean scrollHalfPage(final @NotNull VimEditor editor, final @NotNull VimCaret caret, int rawCount, boolean down) {
-    Editor ijEditor = ((IjVimEditor)editor).getEditor();
-    Caret ijCaret = ((IjVimCaret)caret).getCaret();
-    final CaretModel caretModel = ijEditor.getCaretModel();
-    final int currentLogicalLine = caretModel.getLogicalPosition().line;
-
-    if ((!down && currentLogicalLine <= 0) || (down && currentLogicalLine >= editor.lineCount() - 1)) {
-      return false;
-    }
-
-    final Rectangle visibleArea = getVisibleArea(ijEditor);
-
-    // We want to scroll the screen and keep the caret in the same screen-relative position. Calculate which line will
-    // be at the current caret line and work the offsets out from that
-    int targetCaretVisualLine = getScrollScreenTargetCaretVisualLine(ijEditor, rawCount, down);
-
-    // Scroll at most one screen height
-    final int yInitialCaret = ijEditor.visualLineToY(caretModel.getVisualPosition().line);
-    final int yTargetVisualLine = ijEditor.visualLineToY(targetCaretVisualLine);
-    if (Math.abs(yTargetVisualLine - yInitialCaret) > visibleArea.height) {
-
-      final int yPrevious = visibleArea.y;
-      boolean moved;
-      if (down) {
-        targetCaretVisualLine = getVisualLineAtBottomOfScreen(ijEditor) + 1;
-        moved = scrollVisualLineToTopOfScreen(ijEditor, targetCaretVisualLine);
-      }
-      else {
-        targetCaretVisualLine = getVisualLineAtTopOfScreen(ijEditor) - 1;
-        moved = scrollVisualLineToBottomOfScreen(ijEditor, targetCaretVisualLine);
-      }
-      if (moved) {
-        // We'll keep the caret at the same position, although that might not be the same line offset as previously
-        targetCaretVisualLine = ijEditor.yToVisualLine(yInitialCaret + getVisibleArea(ijEditor).y - yPrevious);
-      }
-    }
-    else {
-      scrollVisualLineToCaretLocation(ijEditor, targetCaretVisualLine);
-
-      final int scrollOffset = getNormalizedScrollOffset(ijEditor);
-      final int visualTop = getVisualLineAtTopOfScreen(ijEditor) + (down ? scrollOffset : 0);
-      final int visualBottom = getVisualLineAtBottomOfScreen(ijEditor) - (down ? 0 : scrollOffset);
-
-      targetCaretVisualLine = max(visualTop, min(visualBottom, targetCaretVisualLine));
-    }
-
-    int logicalLine = EngineEditorHelperKt.visualLineToBufferLine(editor, targetCaretVisualLine);
-    int caretOffset = moveCaretToLineWithStartOfLineOption(editor, logicalLine, caret);
-    caret.moveToOffset(caretOffset);
-
-    return true;
-  }
-
-  // Scrolls current or [count] line to given screen location
-  // In Vim, [count] refers to a file line, so it's a one-based logical line
-  private void scrollLineToScreenLocation(@NotNull Editor editor,
-                                          @NotNull ScreenLocation screenLocation,
-                                          int rawCount,
-                                          boolean start) {
-    final int scrollOffset = getNormalizedScrollOffset(editor);
-
-    int visualLine;
-    if (rawCount == 0) {
-      visualLine = editor.getCaretModel().getVisualPosition().line;
-    }
-    else {
-      final int line = EngineEditorHelperKt.normalizeLine(new IjVimEditor(editor), rawCount - 1);
-      visualLine = new IjVimEditor(editor).bufferLineToVisualLine(line);
-    }
-
-    // This method moves the current (or [count]) line to the specified screen location
-    // Scroll offset is applicable, but scroll jump isn't. Offset is applied to screen lines (visual lines)
-    switch (screenLocation) {
-      case TOP:
-        scrollVisualLineToTopOfScreen(editor, visualLine - scrollOffset);
-        break;
-      case MIDDLE:
-        scrollVisualLineToMiddleOfScreen(editor, visualLine, true);
-        break;
-      case BOTTOM:
-        // Make sure we scroll to an actual line, not virtual space
-        scrollVisualLineToBottomOfScreen(editor, EngineEditorHelperKt.normalizeVisualLine(new IjVimEditor(editor),
-                                                                                          visualLine + scrollOffset));
-        break;
-    }
-
-    if (visualLine != editor.getCaretModel().getVisualPosition().line || start) {
-      int offset;
-      if (start) {
-        offset = moveCaretToLineStartSkipLeading(new IjVimEditor(editor), EngineEditorHelperKt.visualLineToBufferLine(
-          new IjVimEditor(editor), visualLine));
-      }
-      else {
-        offset = moveCaretToLineWithSameColumn(new IjVimEditor(editor),
-                                               EngineEditorHelperKt.visualLineToBufferLine(new IjVimEditor(editor), visualLine),
-                                               new IjVimCaret(editor.getCaretModel().getPrimaryCaret()));
-      }
-
-      new IjVimCaret(editor.getCaretModel().getPrimaryCaret()).moveToOffset(offset);
     }
   }
 
@@ -856,31 +574,5 @@ public class MotionGroup extends VimMotionGroupBase {
 
     final int targetLogicalLine = EngineEditorHelperKt.visualLineToBufferLine(new IjVimEditor(editor), targetVisualLine);
     return moveCaretToLineWithStartOfLineOption(new IjVimEditor(editor), targetLogicalLine, new IjVimCaret(caret));
-  }
-
-  public static class ScrollOptionsChangeListener implements LocalOptionChangeListener<VimDataType> {
-    public static ScrollOptionsChangeListener INSTANCE = new ScrollOptionsChangeListener();
-
-    @Contract(pure = true)
-    private ScrollOptionsChangeListener() {
-    }
-
-    @Override
-    public void processGlobalValueChange(@Nullable VimDataType oldValue) {
-      for (Editor editor : HelperKt.localEditors()) {
-        if (UserDataManager.getVimEditorGroup(editor)) {
-          ScrollViewHelper.scrollCaretIntoView(editor);
-        }
-      }
-    }
-
-    @Override
-    public void processLocalValueChange(@Nullable VimDataType oldValue, @NotNull VimEditor editor) {
-      Editor ijEditor = ((IjVimEditor)editor).getEditor();
-
-      if (UserDataManager.getVimEditorGroup(ijEditor)) {
-        ScrollViewHelper.scrollCaretIntoView(ijEditor);
-      }
-    }
   }
 }
