+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/action/motion/scroll/MotionScrollColumnLeftAction.kt
@@ -28,6 +28,6 @@ class MotionScrollColumnLeftAction : VimActionHandler.SingleExecution() {
     cmd: Command,
     operatorArguments: OperatorArguments,
   ): Boolean {
-    return injector.motion.scrollColumns(editor, cmd.count)
+    return injector.scroll.scrollColumns(editor, cmd.count)
   }
 }
