+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/action/motion/scroll/MotionScrollFirstScreenColumnAction.kt
@@ -23,6 +23,6 @@ class MotionScrollFirstScreenColumnAction : VimActionHandler.SingleExecution() {
     cmd: Command,
     operatorArguments: OperatorArguments,
   ): Boolean {
-    return injector.motion.scrollCaretColumnToDisplayLeftEdge(editor)
+    return injector.scroll.scrollCaretColumnToDisplayLeftEdge(editor)
   }
 }
