+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/action/motion/scroll/MotionScrollHalfWidthRightAction.kt
@@ -43,7 +43,7 @@ class MotionScrollHalfWidthRightAction : VimActionHandler.SingleExecution() {
     operatorArguments: OperatorArguments,
   ): Boolean {
     // Vim's screen width is the full screen width, including columns used for gutters.
-    return injector.motion
+    return injector.scroll
       .scrollColumns(editor, -cmd.count * (injector.engineEditorHelper.getApproximateScreenWidth(editor) / 2))
   }
 }
