+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/action/motion/scroll/MotionScrollLastScreenLineStartAction.kt
@@ -23,6 +23,6 @@ class MotionScrollLastScreenLineStartAction : VimActionHandler.SingleExecution()
   override val flags: EnumSet<CommandFlags> = enumSetOf(CommandFlags.FLAG_IGNORE_SCROLL_JUMP)
 
   override fun execute(editor: VimEditor, context: ExecutionContext, cmd: Command, operatorArguments: OperatorArguments): Boolean {
-    return injector.motion.scrollCurrentLineToDisplayBottom(editor, cmd.rawCount, true)
+    return injector.scroll.scrollCurrentLineToDisplayBottom(editor, cmd.rawCount, true)
   }
 }
