+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/action/motion/scroll/MotionScrollLineAction.kt
@@ -28,7 +28,7 @@ class MotionScrollLineDownAction : VimActionHandler.SingleExecution() {
     cmd: Command,
     operatorArguments: OperatorArguments,
   ): Boolean {
-    return injector.motion.scrollLines(editor, cmd.count)
+    return injector.scroll.scrollLines(editor, cmd.count)
   }
 }
 
@@ -43,6 +43,6 @@ class MotionScrollLineUpAction : VimActionHandler.SingleExecution() {
     cmd: Command,
     operatorArguments: OperatorArguments,
   ): Boolean {
-    return injector.motion.scrollLines(editor, -cmd.count)
+    return injector.scroll.scrollLines(editor, -cmd.count)
   }
 }
